refactor: move `prove_filter` to `filter_base`
