+++ crates/proof-of-sql/src/sql/proof_gadgets/filter_base.rs
@@ -1,7 +1,13 @@
 use crate::{
-    base::{proof::ProofError, scalar::Scalar},
-    sql::proof::{SumcheckSubpolynomialType, VerificationBuilder},
+    base::{database::Column, proof::ProofError, scalar::Scalar, slice_ops},
+    sql::{
+        proof::{FinalRoundBuilder, SumcheckSubpolynomialType, VerificationBuilder},
+        proof_plans::fold_columns,
+    },
 };
+use alloc::{boxed::Box, vec};
+use ark_ff::{One, Zero};
+use bumpalo::Bump;
 
 #[expect(clippy::similar_names)]
 pub(crate) fn verify_filter<S: Scalar>(
@@ -44,4 +50,89 @@ pub(crate) fn verify_filter<S: Scalar>(
     )?;
 
     Ok(())
+}
+
+/// Below are the mappings between the names of the parameters in the math and the code
+/// `c = columns`
+/// `d = filtered_columns`
+/// `n = input_length`
+/// `m = output_length`
+#[expect(clippy::too_many_arguments)]
+#[tracing::instrument(level = "debug", skip_all)]
+pub(crate) fn prove_filter<'a, S: Scalar + 'a>(
+    builder: &mut FinalRoundBuilder<'a, S>,
+    alloc: &'a Bump,
+    alpha: S,
+    beta: S,
+    columns: &[Column<S>],
+    s: &'a [bool],
+    filtered_columns: &[Column<S>],
+    input_length: usize,
+    output_length: usize,
+) {
+    let chi_n = alloc.alloc_slice_fill_copy(input_length, true);
+    let chi_m = alloc.alloc_slice_fill_copy(output_length, true);
+
+    let c_fold = alloc.alloc_slice_fill_copy(input_length, Zero::zero());
+    fold_columns(c_fold, alpha, beta, columns);
+    let d_fold = alloc.alloc_slice_fill_copy(output_length, Zero::zero());
+    fold_columns(d_fold, alpha, beta, filtered_columns);
+
+    let c_star = alloc.alloc_slice_copy(c_fold);
+    slice_ops::add_const::<S, S>(c_star, One::one());
+    slice_ops::batch_inversion(c_star);
+
+    let d_star = alloc.alloc_slice_copy(d_fold);
+    slice_ops::add_const::<S, S>(d_star, One::one());
+    slice_ops::batch_inversion(d_star);
+
+    builder.produce_intermediate_mle(c_star as &[_]);
+    builder.produce_intermediate_mle(d_star as &[_]);
+
+    // sum c_star * s - d_star = 0
+    builder.produce_sumcheck_subpolynomial(
+        SumcheckSubpolynomialType::ZeroSum,
+        vec![
+            (S::one(), vec![Box::new(c_star as &[_]), Box::new(s)]),
+            (-S::one(), vec![Box::new(d_star as &[_])]),
+        ],
+    );
+
+    // c_star + c_fold * c_star - chi_n = 0
+    builder.produce_sumcheck_subpolynomial(
+        SumcheckSubpolynomialType::Identity,
+        vec![
+            (S::one(), vec![Box::new(c_star as &[_])]),
+            (
+                S::one(),
+                vec![Box::new(c_star as &[_]), Box::new(c_fold as &[_])],
+            ),
+            (-S::one(), vec![Box::new(chi_n as &[_])]),
+        ],
+    );
+
+    // d_star + d_fold * d_star - chi_m = 0
+    builder.produce_sumcheck_subpolynomial(
+        SumcheckSubpolynomialType::Identity,
+        vec![
+            (S::one(), vec![Box::new(d_star as &[_])]),
+            (
+                S::one(),
+                vec![Box::new(d_star as &[_]), Box::new(d_fold as &[_])],
+            ),
+            (-S::one(), vec![Box::new(chi_m as &[_])]),
+        ],
+    );
+
+    // d_fold * chi_m - d_fold = 0
+    builder.produce_sumcheck_subpolynomial(
+        SumcheckSubpolynomialType::Identity,
+        vec![
+            (
+                S::one(),
+                vec![Box::new(d_fold as &[_]), Box::new(chi_m as &[_])],
+            ),
+            (-S::one(), vec![Box::new(d_fold as &[_])]),
+        ],
+    );
 }
